+++ apps/pwabuilder-backend/Utils/RequestUtils.cs
@@ -1,8 +1,24 @@
-using PWABuilder.Models;
-
 namespace PWABuilder.Utils
 {
-    public static class RequestUtils
+    public class OutputStatus
+    {
+        public int Status { get; set; }
+        public OutputBody? Body { get; set; }
+    }
+
+    public class OutputBody
+    {
+        public object? Data { get; set; }
+        public OutputError? Error { get; set; }
+    }
+
+    public class OutputError
+    {
+        public required string Object { get; set; }
+        public required string Message { get; set; }
+    }
+
+    public class RequestUtils
     {
         public static OutputStatus CheckParams(HttpRequest req, IEnumerable<string> requiredParams)
         {
